+++ example/dartway_example_flutter/lib/core/router/navigation_zones/app_navigation_zone.dart
@@ -52,6 +52,7 @@ enum AppNavigationZone implements DwNavigationRoute<AppRouterState> {
 
   @override
   List<DwNavigationGuard<AppRouterState>> get zoneGuards => [
-    (state) => !state.isSignedIn ? AuthNavigationZone.auth.fullPath : null,
+    (state, target) =>
+        state.isSignedIn ? null : AuthNavigationZone.signInFrom(target),
   ];
 }